Added compiler flag
  --apply-getters-setters

to enable this change.

Added a new pass ApplyGettersSetters that changes all MemberAccesses
outside of getters, setters, and constructors into function calls so
that

  x.a

becomes

  a(x)

This checkin also adds a fake _methodTokenType argument as the first
parameter to distinguish methods from functions with the same type
signatures so it really becomes

  a(_methodToken, x)

The getter is

  a(_methodTokenType, ClassType)

This code was originally done on the low-level.  This moves this
transformation up to the high-level, but awaits symmetric changes in
the low-level.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@3311 3a8e244f-b0f2-452b-bcba-4c88e055c3ca