Tuples are now records.  This gets rid of lots of special-cased code and is a symmetric change to my recent sequence improvements.  Next is arrays and domains.
Unlike the sequence record, the tuple record is created in the
compiler since it has a different number of types, fields and
functions depending on how many elements are in the tuple.  These
records are still instantiated, they are just created by the compiler
rather than parsed in.

PostAnalysisCleanup, ResolveTypes, and ResolveSymbols now run on all
modules, not just codegened ones.  This is necessary as code in the
prelude may have to be instantiated.

Added typeInfo to FloatLiteral.

Removed ';' from after %token declarations in parser.  This is an
error on the Mac.

Added a new tuple implementation test and two new simple tuple tests.

Removed many tuple futures.

Fixed datablock in prelude and datablock regression.

New regression: arrays/deitz/test_array_integer_3d.

Updated regression file.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@3445 3a8e244f-b0f2-452b-bcba-4c88e055c3ca