Added parsing support for iterators.
Added bool 'yield' field to ReturnStmt to indicate that it is a yield
rather than a return.

Changed bool 'isConstructor' field in FnSymbol to 'fnClass' which can
be any of FN_FUNCTION, FN_CONSTRUCTOR, or FN_ITERATOR.

Added bool 'noparens' field to FnSymbol that is true if the function
was defined without parentheses.

Cleaned up parsing of functions.  There is now only one case for
fndecl rather than three.

Moved check that empty parens are only used on members to a later pass
(stuck it in though this should be part of a preanalysis
checkSemantics pass).

Added sequence return test.

Added a future iterator test.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@3454 3a8e244f-b0f2-452b-bcba-4c88e055c3ca