Added a pass to convert interators into functions that return sequences of whatever they returned before.
The current plan for arrays and domains is to make these classes that
contain a special iterator method.  Then for/forall loops over
sequences, arrays, and domains can be supported by changing them into
while loops over the sequences returned by the special iterator
method.

The longer term approach may be to have a set of iterators for domains
and arrays that can be chosen from past analysis depending on compiler
optimizations.  We'll discuss this with John when he comes back.

Removed iterator future.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@3455 3a8e244f-b0f2-452b-bcba-4c88e055c3ca