Changed the formals of a function from ParamSymbols to DefExprs. Removed the init field of ParamSymbols as it can now use the init field of the DefExpr.
The only symbols not defined in their own DefExprs now are
EnumSymbols.

Added a mechanism to traverse an AList backwards (for_alist_backward).

Cleaned and rewrote the code to insert the default wrapper.  It used
to copy the whole formal list and then remove the formals that didn't
need to go in the wrapper.  It now simply puts the ones that need to
be in in.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@3468 3a8e244f-b0f2-452b-bcba-4c88e055c3ca