Removed DestructureTupleAssignments pass as this is now done before analysis in InsertLiteralTemps.
Removed some uses of TupleType but they are in the current IndexType
so I left them for now.

Update _data record to allocate on construction.

Added a "keep types" pragma for not removing a type from a function
before codegen.  This can be useful in a new version of malloc that
takes a type.  It is used now in the _data constructor.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@3470 3a8e244f-b0f2-452b-bcba-4c88e055c3ca