Simplified and improved the implementation of copy.  Copying a DefStmt, for example, now returns a DefStmt, rather than a Stmt.  The overall structure of copy is similar; the changes are superficial.
One result of this changes is that copy can no longer be called when
an object is NULL.  So c->copy() would now seg fault if c is NULL
rather than work.  To get around this, I added a macro COPY(c) that
can be used when this might be the case.  Alternatively, one can check
for NULL.  There are relatively few uses of this macro now.

Simplified pragma implementation.  They are now a vector of strings in
the BaseAST rather than a BaseAST.  The function hasPragma(char*)
checks to see if the BaseAST has a specific pragma (using strncmp to
compare the start of the pragma with the search string).  The function
addPragma(char*) adds a pragma to a BaseAST's pragma vector.  The
function copyPragmas() is used by copy to copy pragmas on a BaseAST to
the result.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@3472 3a8e244f-b0f2-452b-bcba-4c88e055c3ca