NEW REGRESSION: types/complex/bradc/intfloatelement
Added support for homogeneous tuples as in

  var t : 3*integer;

Parsing support is currently restricted to an integer literal TIMES a
type given by a single identifier.

Added the _chpl_htuple module to implement this as a record.  There
are minimal compiler rewrites:

  number * type_identifier

is changed in the parser to

  _htuple(type_identifier, number)

There are no other rewrites.

Homogeneous tuples do not yet work with standard tuples so you can't
assign one as in

  var t : 3 * integer = (12, 13, 14)

Also because instantiated types can't be fields in a type, these are
implemented without them and the number of components of a homogeneous
tuple is currently limited to 3.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@3487 3a8e244f-b0f2-452b-bcba-4c88e055c3ca