Eliminated the ability to declare multiple variables of the same type by writing, for example,
  var x y z : integer;

In so doing, cleaned up some of the parse code which I hope to
further.

Eliminated the line numbers associated in createEntryPoint pass as
these have little meaning and are somewhat random.

NOTE: Currently

  var x, y, z : integer;

has the meaning that z is of type integer, but x and y have unknown
type.  Their types must be inferred.  This seems already to be the
result of many chapel errors and misunderstandings.  I'd like to
discuss alternatives more fully.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@3631 3a8e244f-b0f2-452b-bcba-4c88e055c3ca