Completed a first version of name mangling scheme. It handles function names, types, and global variables. Local variables will be handled on scope basis. It will not be a problem to add those once the present scheme is tested a bit more.
There are few problems with testing its functionality. A brute
force demangling scheme works to the point that it does
interfere with some hidden convention used for cnames (uncomment)
the indicated code to try it out. Therefore I think Steve should
take a look and maybe we can come up with a way to slowly
integrate the new scheme once we are sure that it does what
we want. More functionality needs to and will be added.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@3689 3a8e244f-b0f2-452b-bcba-4c88e055c3ca