REGRESSIONS UPDATE NOTE:   Trading 2 regressions for 1: trivial/mjoyner/seq/seq_reverse,   types/seq/deitz/test_seq_reverse1 now work;   classes/deitz/test_nested_record2 now fails.
Reworked the sequence implementation so that appending, prepending,
and concatenating copy the sequence.  Eventually we'll want to
optimize this, but for the time being this is a safer implementation.

Removed two futures.

Added two new tests, to test a tuple with an inferred type and to test
a clone function in an instantiated class that returns the constructor
using the same type variable.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@3749 3a8e244f-b0f2-452b-bcba-4c88e055c3ca