Removed Expr methods intVal and precedence as these will need to be substantively revisited after BinOp, UnOp, AssignOp, and ParenOpExpr are unified into CallExpr.
Futurized test/types/enum/deitz/test_enum5 because this test
initializes an enum symbol with an expression, in this case, -1.  Do
we need a way to compute this value?  Should we require that all enum
symbols be 'param' integers?  This is not so easy to compute since
unary minus can be overloaded.  In general, we do need a way to fold
constants and params.  How should we do this?


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@3774 3a8e244f-b0f2-452b-bcba-4c88e055c3ca