Fixed the _chpl_malloc bug that Shannon tracked down and removed the #if that John put in.
The problem was that the sizeof was now being taken of the pointer,
not the struct.  In the un #defed code, an underscore was put in front
of the name of the type, as that is what we want to take the sizeof
for malloc purposes.  I put it back in.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@3853 3a8e244f-b0f2-452b-bcba-4c88e055c3ca