Removed Tuple (the Expr subclass for Tuple literals).  This was removed early enough to just get rid of it.
Removed code for destructuring tuples in analysis.  This has not been
used for a while as destructuring is done in the high-level.

Made a fix to analysis/ast.cpp to keep two tests from starting to
seg-fault.  John, please see my comment in analysis/ast.cpp and make
sure this is an okay change.  If not, we can futurize these tests if
something deeper is going on.

Added a test case to test '(i, j) = (j, i)' which swaps two integers.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@3865 3a8e244f-b0f2-452b-bcba-4c88e055c3ca