Added  Vec<Type*> Type::typeParents  Vec<Type*> Type::dispatchParents to represent the type hierarchies.
Removed
 Type* Type::parentType
 Vec<Type*> ClassType::parentClasses
 Type* ClassType::parentStruct

Since most of this wasn't used, it was a straightforward change.  It's
not set up very well as the one current usage seems strange.  I
commented out a use of one that was never set in analysis.  John, can
you take a look at how this might be rearranged and doublecheck the
current usage in resolveTypes.cpp.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@3920 3a8e244f-b0f2-452b-bcba-4c88e055c3ca