Simplified parsing and construction of ForLoopStmt and ForallExpr, prior a mess, in a successful effort to remove the last AList of symbols or types.  Exception: ModuleSymbols are in an AList.
Added verification to keep symbols and types from being in ALists.
Unlike Exprs and Stmts, these can be used multiple times in the
program structure so it is best to keep them in Vecs.

I also added verification to make sure Type is never used (only
subclasses of it) now that John built up the PrimitiveType.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@3932 3a8e244f-b0f2-452b-bcba-4c88e055c3ca