Changed LabelStmt so that the field
  LabelSymbol* label

is now

  DefExpr* defLabel

Removed special-cased code to handle this symbol definition.  It is
now handled with all the other DefExpr symbol definitions.

As a side effect, this change fixes a bug (that didn't yet show up) in
which if a label were in a cloned function, the label symbol would not
be cloned.  This could have been fixed without this change, but the
uniformity is a better way of making this fix.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4009 3a8e244f-b0f2-452b-bcba-4c88e055c3ca