Added support for mutually recursive nested functions.  Removed a mutually recursive future test and added an additional one.
Also added a small change to inline function traversal to ensure that pragmas can be gathered from the function symbol.  Sometimes they are attached to the function symbol and sometimes they are attached to the function definition statement. Now the inline function traversal only has to look in one place to determine if a function should be inlined.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@4011 3a8e244f-b0f2-452b-bcba-4c88e055c3ca